+++ modules_k/nat_traversal/nat_traversal.c
@@ -1527,7 +1527,8 @@ send_keepalive(NAT_Contact *contact)
                    keepalive_params.extra_headers);
 
     if (len >= sizeof(buffer)) {
-        LM_ERR("keepalive message is longer than %d bytes\n", sizeof(buffer));
+        LM_ERR("keepalive message is longer than %lu bytes\n",
+			(unsigned long)sizeof(buffer));
         return;
     }
